+++ packages/plugin-zod/src/generators/__snapshots__/enumBooleanLiteral.ts
@@ -1,0 +1,3 @@
+import { z } from "zod";
+
+ export const enumBooleanLiteral = z.union([z.literal(true), z.literal(false)]);